Moved some functions to separate files
