+++ .gitignore
@@ -18,3 +18,4 @@ cache
 node_modules
 dist
 merge
+todo.md
