+++ src/supplement/index.ts
@@ -1,0 +1,17 @@
+export class Coordinate {
+	x: number;
+	y: number;
+
+	constructor(x = 0, y = 0) {
+		[this.x, this.y] = [x, y];
+	}
+
+	setCoordinates(x?: number, y?: number) {
+		this.x = x ?? this.x;
+		this.y = y ?? this.y;
+	}
+
+	reset() {
+		[this.x, this.y] = [0, 0];
+	}
+}
